+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 #### Next Release
 - Added Query.first()
 - Added Query.firstOrDefault()
+- Added Query.toList(converter)
 
 #### v0.1.0 (2019-11-22)
 - Initial release